Move Coordinate spaces to their own file, and clean up universe stuff
* Renamed Vector3s to Vector3g, since Vector3s also refers to a Vector3
  that uses shorts
